--- src/ehrdata/core/_compat.py
@@ -1,9 +1,0 @@
-def lazy_import_torch() -> None:
-    try:
-        import torch
-
-        return torch
-    except ImportError:
-        raise ImportError(
-            "The optional module 'torch' is not installed. Please install it using 'pip install ehrdata[torch]'."
-        ) from None
